# chore(test): add micro-benchmark performance test for toPromise operator