--- a/perf/micro/immediate-scheduler/operators/topromise-ctor.js
+++ b/perf/micro/immediate-scheduler/operators/topromise-ctor.js
@@ -1,0 +1,18 @@
+var RxOld = require("rx");
+var RxNew = require("../../../../index");
+
+module.exports = function (suite) {
+
+  var oldToPromiseWithImmediateScheduler = RxOld.Observable.of(25, RxOld.Scheduler.immediate).toPromise(Promise);
+  var newToPromiseWithImmediateScheduler = RxNew.Observable.of(25).toPromise(Promise);
+
+  return suite
+    .add('old toPromise(ctor) with immediate scheduler', function () {
+      oldToPromiseWithImmediateScheduler.then(_then);
+    })
+    .add('new toPromise(ctor) with immediate scheduler', function () {
+      newToPromiseWithImmediateScheduler.then(_then);
+    });
+  
+  function _then(x) { }
+};