--- a/perf/micro/immediate-scheduler/operators/topromise.js
+++ b/perf/micro/immediate-scheduler/operators/topromise.js
@@ -1,0 +1,18 @@
+var RxOld = require("rx");
+var RxNew = require("../../../../index");
+
+module.exports = function (suite) {
+
+  var oldToPromiseWithImmediateScheduler = RxOld.Observable.of(25, RxOld.Scheduler.immediate).toPromise();
+  var newToPromiseWithImmediateScheduler = RxNew.Observable.of(25).toPromise();
+
+  return suite
+    .add('old toPromise() with immediate scheduler', function () {
+      oldToPromiseWithImmediateScheduler.then(_then);
+    })
+    .add('new toPromise() with immediate scheduler', function () {
+      newToPromiseWithImmediateScheduler.then(_then);
+    });
+  
+  function _then(x) { }
+};