@blesh 2015-09-22
@kwonoj  it makes sense this one would be close, and perhaps a little slower. It looks like I'm introducing a little more closure and another assertion. I think we can trim the fat here.