@kwonoj 2015-09-22
Reason I was suspicious about my test case design was test results was vary and only some of system shows slower results than original. I was able to see some of additional checks as you mentioned, bit unsure about results is not consistent over system configuration. Still it's interesting topic to track down further to see where does difference comes.