@blesh 2015-09-22
Another thing to pay attention to, is that % accuracy at the end of the test.

> old toPromise() with immediate scheduler x 351,507 ops/sec **±4.64%** (81 runs sampled)
> new toPromise() with immediate scheduler x 322,127 ops/sec **±9.00%** (65 runs sampled)
-8.36% slower than Rx2

The "old" toPromise test there could be off by up to around 16,000 ops per sec, and the "new" toPromise could be off by up to almost 29,000 ops per second.

When I ran it, RxJS Next was "2% faster" and I had slightly higher accuracy.  I suspect these two are just going to be on par, because the design of Promise pretty much means you need to introduce some closure.