@blesh 2015-09-22
> old toPromise() with immediate scheduler x 298,514 ops/sec ±4.15% (82 runs sampled)
> new toPromise() with immediate scheduler x 306,881 ops/sec ±3.29% (83 runs sampled)
> 	2.8% faster than Rx2

> old toPromise() with immediate scheduler x 309,282 ops/sec ±4.89% (77 runs sampled)
> new toPromise() with immediate scheduler x 330,398 ops/sec ±3.46% (83 runs sampled)
> 	6.83% faster than Rx2

A few runs on my machine and I'm unable to get to to lean the other way.