@blesh 2015-09-22
merged with 5e8be38

Thanks, @kwonoj!!